Fixed transition duration...

+++ AssettoServer/Server/Preset/PresetData.cs
@@ -4,7 +4,7 @@ public class PresetData
 {
     public PresetType? Type { get; set; }
     public PresetType? UpcomingType { get; set; }
-    public double TransitionDuration { get; set; }
+    public int TransitionDuration { get; set; }
     public bool IsInit { get; set; } = false;
 
     public PresetData(PresetType? type, PresetType? upcomingType)
